requirements.txt
flask==1.1.2
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask-script==2.0.6
gunicorn==20.0.4
psycopg2==2.8.5
flask-dotenv==0.1.2
